exe2/requirements.txt
mysql-connector-python
pandas
pyarrow